Edited structure for service
