--- server/Auth service/JWT/Token.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
