--- server/Auth service/authResponse.py
@@ -1,4 +1,0 @@
-class authResponse(dict):
-    def __init__(self, token, expiresin):
-        self.token = token.decode('utf-8')
-        self.expiresin = expiresin
