+++ server/Auth service/config.py
@@ -1,0 +1,4 @@
+SECRET_KEY = "adbd35fdfb55f0f0a431b7788707256dbd1e69be93f26aafb1d69317e3a613ef"
+ALGORITHM = "HS256"
+TIME_DELTA_FOR_ACCESS = 15
+TIME_DELTA_FOR_REFRESH = 30
